% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_geomean.R
\name{calc_geomean}
\alias{calc_geomean}
\title{Calculate Geometric Mean}
\usage{
calc_geomean(x, ci = NA, R = 1000)
}
\arguments{
\item{x}{a vector of numeric values}

\item{ci}{numeric value for the confidence interval. Defaults to NA.
Confidence intervals are calculated using bootstrap percentile intervals.}

\item{R}{number of bootstrap resamples, defaults to 1000.}
}
\value{
A numeric value, the geometric mean of x. Or a named list with
geometric mean and the confidence intervals.
}
\description{
Calculates the geometric mean of x and, optionally, the confidence interval.
}
\details{
\code{x} must be a vector of positive numeric values. The geometric
mean is calculated as:

\loadmathjax

\mjsdeqn{\bar{x} = \exp  \left[\frac{1}{n}\sum_{i=1}^{n}\log(x_i)\right] = e^\bar{y} }

where:

\mjsdeqn{\bar{y} = \frac{1}{n} \sum_{i=1}^n y_i}

\mjsdeqn{y_i = log(x_i), \;\; i = 1, 2, \ldots, n}

or more simply, the geometric mean of \code{x} is equal to the antilog of the mean of log(\code{x}).
}
\examples{
x <- rlnorm(1000, log(100), log(3))
calc_geomean(x)
calc_geomean(x, ci = .95, R = 1000)
}
