% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_NLCD.R
\name{extract_NLCD}
\alias{extract_NLCD}
\title{Extract NLCD Areas}
\usage{
extract_NLCD(
  feature,
  id_variable,
  area_units = "acres",
  crs,
  year = 2016,
  dataset = "Land_Cover"
)
}
\arguments{
\item{feature}{feature to extract and calculate NLCD class areas. Can be class sp or sf with one or multiple features.}

\item{id_variable}{unquoted id variable in feature, must have unique value for each row.}

\item{area_units}{desired output units. probably one of \code{c("acres","mi^2")}. See \code{units::valid_udunits()} for more units.}

\item{crs}{coordinate reference system: integer with the EPSG code, or character with proj4string}

\item{year}{An integer representing the year of desired NLCD product. Acceptable values are 2016 (default), 2011, 2008 (landcover only), 2006, 2004, and 2001.}

\item{dataset}{A character string representing type of the NLCD product. Acceptable values are 'Impervious', 'Land_Cover', 'Tree_Canopy' (2011 and 2016 only),}
}
\value{
something
}
\description{
Extracts and calculates the area of NLCD classes for a specified feature. Returns a long tibble with percent and area for each class and feature.
}
