% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nlcd_legend}
\alias{nlcd_legend}
\title{NLCD Legend}
\format{
A tibblee with two variables
\describe{
\item{class}{NLCD class name}
\item{value}{NLCD class value}
...
}
}
\source{
\url{hhttps://www.mrlc.gov/data/legends/national-land-cover-database-2016-nlcd2016-legend}
}
\usage{
nlcd_legend
}
\description{
Tibble with value and label for NLCD classes.
}
\keyword{datasets}
