% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_xl.R
\name{round_xl}
\alias{round_xl}
\title{Inelegant Excel Rounding}
\usage{
round_xl(x, digits = 0)
}
\arguments{
\item{x}{a numeric value}

\item{digits}{a numeric value indicating the desired digits of precision}
}
\value{
a rounded numeric value
}
\description{
This function mimics the rounding found in Excel where 0.5 is always round
away from zero. It is generally recommended to avoid using this, but
sometimes we have to mimic calculations done inside an Excel workbook.
}
\examples{
x <- .5 + -2:4
# using base round function
# expected values are: -2 0 0 2 2 4 4
round(x)

# using Excel rounding
# the expected values are: -2 -1  1  2  3  4  5
round_xl(x)
}
\seealso{
\code{\link[=round]{round()}} for base rounding and details of rounding at 5.
}
